Feature/user

+++ src/main/java/com/hb/acadia/service/RoleService.java
@@ -2,11 +2,18 @@ package com.hb.acadia.service;
 
 import java.util.List;
 
+import javax.transaction.Transactional;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.web.access.ExceptionTranslationFilter;
 import org.springframework.stereotype.Service;
+import org.springframework.web.servlet.mvc.method.annotation.ExceptionHandlerExceptionResolver;
 
 import com.hb.acadia.model.user.Role;
+import com.hb.acadia.model.user.User;
 import com.hb.acadia.repository.RoleRepository;
+
+import lombok.extern.slf4j.Slf4j;
 
 /**
  * Service class
@@ -14,11 +21,24 @@ import com.hb.acadia.repository.RoleRepository;
  * @author simonaliotti
  *
  */
+@Slf4j
 @Service
 public class RoleService {
 
 	@Autowired
-	RoleRepository roleRepository;
+	private RoleRepository roleRepository;
+	@Autowired
+	private UserService userService;
+
+	/**
+	 * Method to get a role by roleName
+	 * 
+	 * @param roleName
+	 * @return the given role
+	 */
+	public Role getRoleByRoleName(String roleName) {
+		return roleRepository.findRoleByRoleName(roleName);
+	}
 
 	/**
 	 * Method to get all roles
@@ -33,24 +53,40 @@ public class RoleService {
 	 * Method to create role
 	 * 
 	 * @param role
+	 * @return the created Role
 	 */
+	@Transactional
 	public Role createRole(Role role) {
 		return roleRepository.save(role);
+
 	}
 
 	/**
-	 * Delete a role 
+	 * Delete a role Delete all roles. Set users's role to null before deleting role
+	 * 
 	 * @param role
+	 * @throws IllegalAccessException
 	 */
-	public void deleteRole(Role role) {
-		roleRepository.delete(role);
+	@Transactional
+	public void deleteRole(Role role) throws IllegalAccessException {
+		List<User> users = userService.getUserByRole(role);
+		if (users.isEmpty()) {
+			roleRepository.delete(role);
+		} else {
+			throw new IllegalAccessException();
+		}
+
 	}
 
 	/**
-	 * Delete all roles. 
-	 *  **** WARNINGS **** Be shure that no user are connected to the roles
+	 * Delete all roles. the roles
+	 * 
+	 * @throws IllegalAccessException
 	 */
-	public void deleteAll() {
-		roleRepository.deleteAll();
+	@Transactional
+	public void deleteAll() throws IllegalAccessException {
+		for (Role role : getRoles()) {
+			deleteRole(role);
+		}
 	}
 }
